requirements.txt
asgiref==3.5.2
Django==4.1.2
sqlparse==0.4.3
tzdata==2022.4
certifi==2022.6.15
charset-normalizer==2.0.12
django-admin-thumbnails==0.2.6
django-bootstrap-form==3.4
django-bootstrap-modal-forms==2.2.0
django-ckeditor==6.4.2
django-filter==22.1
django-globals==0.3.2
django-jalali==6.0.0
django-js-asset==2.0.0
django-phone-field==1.8.1
django-taggit==3.0.0
ghasedakpack==0.1.9
idna==3.3
jdatetime==4.1.0
Pillow==9.1.1
pytz==2022.1
requests==2.28.0
six==1.16.0
sorl-thumbnail==12.8.0
urllib3==1.26.9
